Update test_system_tables_lazy_load to cover both cases and make it run
