+++ tests/integration/test_system_tables_lazy_load/configs/config_lazy.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0"?>
+<yandex>
+    <system_tables_lazy_load>true</system_tables_lazy_load>
+</yandex>
