--- tests/integration/test_system_tables_non_lazy_load/configs/config_non_lazy.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0"?>
-<yandex>
-    <system_tables_lazy_load>false</system_tables_lazy_load>
-</yandex>
